backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-multipart==0.0.9
openai==1.35.3
python-dotenv==1.0.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
pandas==2.2.2
openpyxl==3.1.2